lint

+++ flyteplugins/go/tasks/plugins/webapi/bigquery/config_flags_test.go
@@ -133,7 +133,7 @@ func TestConfig_SetFlags(t *testing.T) {
 			testValue := "1"
 
 			cmdFlags.Set("webApi.writeRateLimiter.qps", testValue)
-			if vInt, err := cmdFlags.GetInt("webApi.writeRateLimiter.qps"); err == nil {
+			if vInt, err := cmdFlags.GetFloat64("webApi.writeRateLimiter.qps"); err == nil {
 				testDecodeJson_Config(t, fmt.Sprintf("%v", vInt), &actual.WebAPI.WriteRateLimiter.QPS)
 
 			} else {
